Add goal history detail screen
