+++ lib/screens/goals_overview_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import '../services/goals_service.dart';
+import 'goal_history_screen.dart';
 
 class GoalsOverviewScreen extends StatelessWidget {
   const GoalsOverviewScreen({super.key});
@@ -28,19 +29,28 @@ class GoalsOverviewScreen extends StatelessWidget {
           final g = goals[index];
           final progress = (g.progress / g.target).clamp(0.0, 1.0);
           final completed = g.completedAt != null;
-          return Container(
-            margin: const EdgeInsets.only(bottom: 12),
-            padding: const EdgeInsets.all(12),
-            decoration: BoxDecoration(
-              color: Colors.grey[850],
-              borderRadius: BorderRadius.circular(8),
-            ),
-            child: Row(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                if (g.icon != null) ...[
-                  Icon(g.icon, color: accent),
-                  const SizedBox(width: 12),
+          return InkWell(
+            onTap: completed
+                ? () => Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                          builder: (_) => GoalHistoryScreen(index: index)),
+                    )
+                : null,
+            borderRadius: BorderRadius.circular(8),
+            child: Container(
+              margin: const EdgeInsets.only(bottom: 12),
+              padding: const EdgeInsets.all(12),
+              decoration: BoxDecoration(
+                color: Colors.grey[850],
+                borderRadius: BorderRadius.circular(8),
+              ),
+              child: Row(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  if (g.icon != null) ...[
+                    Icon(g.icon, color: accent),
+                    const SizedBox(width: 12),
                 ],
                 Expanded(
                   child: Column(
@@ -80,6 +90,7 @@ class GoalsOverviewScreen extends StatelessWidget {
                   color: completed ? Colors.green : Colors.grey,
                 ),
               ],
+              ),
             ),
           );
         },
